requirements.txt
playwright==1.52.0
python-dotenv==1.1.0
pyyaml==6.0.2
